+++ src/main.rs
@@ -3,6 +3,7 @@
 mod app;
 mod config;
 mod doorbell;
+mod initializer;
 mod mailroom;
 mod pane;
 mod plugins;
@@ -13,6 +14,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         Some("send") => doorbell::send_command(),
         Some("control") => doorbell::control_command(),
         Some("pulse") => doorbell::pulse_command(),
+        Some("init") => initializer::command(),
         Some("plugin") => plugins::command(),
         Some("toolbox") => toolbox::command(),
         _ => app::run(),